Fixed comments, refactored routes
