+++ moninag/registration/urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-from registration.views import activate, inactive_account, login, auth_view, logout, profile, invalid_login, register_user, register_success
+from registration.views import activate, inactive_account, login, auth, logout, profile, invalid_login, register_user, register_success
 
 # Uncomment the next two lines to enable the admin:
 # from django.contrib import admin
@@ -7,14 +7,14 @@ from registration.views import activate, inactive_account, login, auth_view, log
 
 urlpatterns = [
     # auth
-    url(r'^$', login, name='login'),
-    url(r'^auth/$', auth_view, name='auth_view'),
+    url(r'^$', auth, name='auth'),
+    url(r'^login/$', login, name='login'),
     url(r'^logout/$', logout, name='logout'),
     url(r'^profile/$', profile, name='profile'),
     url(r'^invalid/$', invalid_login, name='invalid_login'),
     url(r'^inactive_account/$', inactive_account, name='inactive_account'),
     url(r'^activate/(?P<activation_key>[0-9]+)/$', activate, name='activate'),
-    url(r'^register/$', register_user, name='register_user'),
+    url(r'^register_user/$', register_user, name='register_user'),
     url(r'^register_success/$', register_success, name='register_success'),
