--- moninag/static/src/ts/app/registration/_services/authentication.service.ts
@@ -1,16 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Http, Headers, Response, RequestOptions } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map'
-
-@Injectable()
-export class AuthenticationService {
-    constructor(private http: Http) { }
-
-    login(email: string, password: string) {
-        let headers = new Headers({ 'Content-Type': 'application/json', 'HTTP_X_REQUESTED_WITH': 'XMLHttpRequest'});
-        let options = new RequestOptions({ headers: headers });
-        let body = JSON.stringify({'email': email, 'password': password});
-        return this.http.post('/accounts/auth/', body, options).map((res: Response) => res.json());
-    }
-}
