--- moninag/static/src/ts/app/registration/_services/index.ts
@@ -1,3 +1,0 @@
-export * from './alert.service';
-export * from './authentication.service';
-export * from './user.service';
