--- moninag/static/src/ts/app/registration/login/login.component.ts
@@ -1,39 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
-
-import { AlertService, AuthenticationService } from '../_services/index';
-
-@Component({
-    selector: 'login',
-    templateUrl: 'static/src/ts/app/registration/login/login.component.html'
-})
-
-export class LoginComponent {
-    model: any = {};
-    loading = false;
-
-    constructor(
-        private route: ActivatedRoute,
-        private router: Router,
-        private authenticationService: AuthenticationService,
-        private alertService: AlertService) { }
-
-    login() {
-        this.loading = true;
-        this.authenticationService.login(this.model.email, this.model.password)
-            .subscribe(
-                data => {
-                    if(data.success) {
-                        window.location.href = '/accounts/profile/';
-                    } else {
-                        this.alertService.error(data.error);
-                        this.loading = false;
-                    }
-                },
-                error => {
-                    this.alertService.error(error);
-                    this.loading = false;
-                }
-                );
-    }
-}
