--- moninag/static/src/ts/app/registration/message/message.component.html
@@ -1,1 +1,0 @@
-<div *ngIf="message" [ngClass]="{ 'message': message, 'alert-success': message.type === 'success', 'alert-danger': message.type === 'error' }">{{message.text}}</div>
