+++ moninag/static/src/ts/auth/_services/alert.service.ts
@@ -1,0 +1,37 @@
+import { Injectable } from '@angular/core';
+import { Router, NavigationStart, NavigationEnd } from '@angular/router';
+import { Observable } from 'rxjs';
+import { Subject } from 'rxjs/Subject';
+
+@Injectable()
+export class AlertService {
+
+    private subject = new Subject<any>();
+    private keepAfterNavigationChange = false;
+
+    constructor(private router: Router) {
+        // clear alert message on route change
+        router.events.subscribe(event => {
+
+            if (event instanceof NavigationEnd) {
+                // clear alert
+                this.subject.next();
+            }
+        });
+    }
+
+
+    success(message: string, keepAfterNavigationChange = false) {
+        this.keepAfterNavigationChange = keepAfterNavigationChange;
+        this.subject.next({ type: 'success', text: message });
+    }
+
+    error(message: string, keepAfterNavigationChange = false) {
+        this.keepAfterNavigationChange = keepAfterNavigationChange;
+        this.subject.next({ type: 'error', text: message });
+    }
+
+    getMessage(): Observable<any> {
+        return this.subject.asObservable();
+    }
+}
