+++ moninag/static/src/ts/auth/login/login.component.html
@@ -1,0 +1,22 @@
+<div class="col-md-4 col-md-offset-4">
+    <h2>Sign In</h2>
+    <form name="form" (ngSubmit)="f.form.valid && login()" #f="ngForm" novalidate>
+        <div class="form-group" [ngClass]="{ 'has-error': f.submitted && !email.valid }">
+            <label for="email">Email</label>
+            <input type="email" class="form-control" name="email" [(ngModel)]="model.email" #email="ngModel" required />
+            <div *ngIf="f.submitted && !email.valid" class="help-block">Email is required</div>
+        </div>
+        <div class="form-group" [ngClass]="{ 'has-error': f.submitted && !password.valid }">
+            <label for="password">Password</label>
+            <input type="password" class="form-control" name="password" [(ngModel)]="model.password" #password="ngModel" required />
+            <div *ngIf="f.submitted && !password.valid" class="help-block">Password is required</div>
+        </div>
+        <div class="form-group">
+            <button [disabled]="loading" class="btn btn-primary">Sign In</button>
+            <div class="pull-right">
+                Don't have an account yet?
+                <a [routerLink]="['/accounts/register_user']" class="btn btn-link">Sign Up</a>
+            </div>
+        </div>
+    </form>
+</div>
