+++ moninag/static/src/ts/auth/main.ts
@@ -1,0 +1,5 @@
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+
+import { AuthModule } from './auth.module';
+
+platformBrowserDynamic().bootstrapModule(AuthModule);
