+++ moninag/static/src/ts/auth/message/message.component.html
@@ -1,0 +1,1 @@
+<div *ngIf="message" [ngClass]="{ 'message': message, 'alert-success': message.type === 'success', 'alert-danger': message.type === 'error' }">{{message.text}}</div>
