+++ moninag/templates/registration/already_activated.html
@@ -1,9 +1,9 @@
-{% extends "base.html" %}
+{% extends "../index.html" %}
 
 {% block content %}
 
   <h2>This account has been already activated!</h2>
 
-  <p>Click <a href="{% url 'login' %}">here</a> to sign in.</p>
+  <p>Click <a href="{% url 'auth' %}">here</a> to sign in.</p>
 
 {% endblock %}
