--- moninag/templates/registration/base.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <base href="/">
-    <meta charset="utf-8">
-    <title>Moninag</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-    <link rel="stylesheet" type="text/css" href="static/css/styles.css">
-
-    <script src="static/lib/Rx.js"></script>
-    <script src="static/lib/zone.js"></script>
-    <script src="static/lib/Reflect.js"></script>
-
-    <script async="" src="static/js/source_app.js"></script>
-</head>
-<body>
-    {% block content %}
-
-    {% endblock %}
-</body>
-</html>
