+++ moninag/templates/registration/inactive_account.html
@@ -1,8 +1,8 @@
-{% extends "base.html" %}
+{% extends "../index.html" %};
 
 {% block content %}
 
   <h2>Please activate your account before login!</h2>
-  <p>Click <a href="{% url 'login' %}">here</a> to sign in again.</p>
+  <p>Click <a href="{% url 'auth' %}">here</a> to sign in again.</p>
 
 {% endblock %}
