+++ moninag/templates/registration/invalid_login.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "../index.html" %}
 
 {% block content %}
