+++ moninag/templates/registration/login.html
@@ -1,5 +1,8 @@
-{% extends "base.html" %}
+{% extends "../index.html" %}
 
+{% block jsApp %}
+    <script async="" src="static/js/auth_app.js"></script>
+{% endblock %}
 {% block content %}
-  <moninag-app></moninag-app>
+    <auth-app></auth-app>
 {% endblock %}
