+++ moninag/templates/registration/logout.html
@@ -1,8 +1,8 @@
-{% extends "base.html" %}
+{% extends "../index.html" %}
 
 {% block content %}
 
   <h2>Logged out!</h2>
-  <p>Click <a href="{% url 'login' %}">here</a> to sign in again</p>
+  <p>Click <a href="{% url 'auth' %}">here</a> to sign in again</p>
 
 {% endblock %}
