+++ moninag/templates/registration/register_success.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "../index.html" %}
 
 {% block content %}
